Changed TenorProviderTestUtils to TenorProviderTestFixtures, also refactored the class to become a object
